Implement policy in controller

+++ app/Http/Controllers/QuestionsController.php
@@ -62,6 +62,7 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        $this->authorize('update', $question);
         return view("questions.edit", compact('question'));
     }
 
@@ -74,6 +75,7 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        $this->authorize('update', $question);
         $question->update($request->only('title', 'body'));
         return redirect('/questions')->with('success', "Your question has been updated.");
     }
@@ -86,6 +88,7 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        $this->authorize('delete', $question);
         $question->delete();
         return redirect('/questions')->with('success', "Your question has been deleted.");
     }
